+++ domain/src/test/java/woowacourse/movie/domain/MovieTicketTest.kt
@@ -1,9 +1,7 @@
-package domain
+package woowacourse.movie.domain
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import woowacourse.movie.domain.MovieTicket
-import woowacourse.movie.domain.PeopleCount
 import java.time.LocalDateTime
 
 class MovieTicketTest {
